--- src/checkers/youtube.ts
@@ -1,19 +1,0 @@
-import { Shortener, MyURL } from './checker';
-
-export class Youtube implements Shortener {
-  canShorten(url: MyURL): boolean {
-    return (
-      (url.hostname.endsWith('youtube.com') ||
-        url.hostname.endsWith('youtu.be')) &&
-      url.pathname === '/watch' &&
-      url.searchParams.has('v')
-    );
-  }
-
-  shorten(url: MyURL): MyURL {
-    url.host = 'youtu.be';
-    url.pathname = url.searchParams.get('v') || '';
-    url.deleteSearch();
-    return url;
-  }
-}
